# docs/swagger.yaml
basePath: /api/v1
definitions:
  v1.errorResponse:
    properties:
      message:
        type: string
    type: object
  v1.staffSignInInput:
    properties:
      email:
        maxLength: 64
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  v1.staffSignUpInput:
    properties:
      email:
        maxLength: 64
        type: string
      name:
        maxLength: 64
        minLength: 2
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
      phone:
        maxLength: 10
        type: string
    required:
    - email
    - name
    - password
    - phone
    type: object
  v1.staffUpdateInfoInput:
    properties:
      name:
        maxLength: 64
        minLength: 2
        type: string
      phone:
        maxLength: 10
        type: string
    required:
    - name
    - phone
    type: object
  v1.userSignInInput:
    properties:
      email:
        maxLength: 64
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  v1.userSignUpInput:
    properties:
      email:
        maxLength: 64
        type: string
      name:
        maxLength: 64
        minLength: 2
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
      phone:
        maxLength: 10
        type: string
    required:
    - email
    - name
    - password
    - phone
    type: object
  v1.userUpdateInfoInput:
    properties:
      name:
        maxLength: 64
        minLength: 2
        type: string
      phone:
        maxLength: 10
        type: string
    required:
    - name
    - phone
    type: object
host: localhost:8000
info:
  contact:
    email: amankostleukulov@gmail.com
    name: Aman
  description: REST API for Auction App
  title: Auction API
  version: "1.0"
paths:
  /auth-staff/sign-in:
    post:
      consumes:
      - application/json
      description: Sign in to staff account
      parameters:
      - description: Sign In
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.staffSignInInput'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Sign in
      tags:
      - Staff
  /auth-staff/sign-up:
    post:
      consumes:
      - application/json
      description: Owner creates staff
      parameters:
      - description: Sign Up
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.staffSignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Sign up
      tags:
      - Staff
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Sign in to user account
      parameters:
      - description: Sign In
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.userSignInInput'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Sign in
      tags:
      - Account
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Create account with role user
      parameters:
      - description: Sign Up
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.userSignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Sign up
      tags:
      - Account
  /staff/info:
    get:
      consumes:
      - application/json
      description: Get staff information
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Staff Info
      tags:
      - Staff
  /staff/info/{uuid}:
    post:
      consumes:
      - application/json
      description: Update staff information
      parameters:
      - description: Staff uuid
        in: path
        name: uuid
        required: true
        type: string
      - description: Update Staff Information
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.staffUpdateInfoInput'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Update Staff Info
      tags:
      - Staff
  /user/info:
    get:
      consumes:
      - application/json
      description: Get user information
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: User Info
      tags:
      - Account
  /user/info/{uuid}:
    post:
      consumes:
      - application/json
      description: Update user information
      parameters:
      - description: User uuid
        in: path
        name: uuid
        required: true
        type: string
      - description: Update User Information
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.userUpdateInfoInput'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Update User Info
      tags:
      - Account
swagger: "2.0"